backend/requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
Authlib==1.5.2
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.8
cryptography==44.0.2
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi-cli==0.0.7
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
icecream==2.1.4
idna==3.10
iniconfig==2.1.0
ipython_pygments_lexers==1.1.1
ipython==9.0.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
kombu==5.5.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
packaging==24.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic_core==2.33.1
pydantic==2.11.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich-toolkit==0.14.1
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy-Utils==0.41.2
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1